Worked on the Service Validators

+++ DNBarbershop.Core/Services/BarberService.cs
@@ -24,51 +24,95 @@ namespace DNBarbershop.Core.Services
             {
                 return false;
             }
-         return true;
+            return true;
         }
         public async Task Add(Barber barber)
         {
-            
-            await _barberRepository.Add(barber);
+            if (ValidateBarber(barber))
+            {
+                await _barberRepository.Add(barber);
+            }
+            else
+            {
+                throw new ArgumentException("Validation didn't pass.");
+            }
         }
-
         public async Task Delete(Guid id)
         {
-            await _barberRepository.Delete(id);
+            if (BarberValidator.BarberExists(id))
+            {
+                await _barberRepository.Delete(id);
+            }
+            else
+            {
+                throw new ArgumentException("This barber doesn't exist.");
+            }
         }
-
         public async Task DeleteAll()
         {
-            await _barberRepository.DeleteAll();
+            if (await _barberRepository.GetCount()<=0)
+            {
+                throw new ArgumentException("Nothing to delete here.");
+            }
+            else
+            {
+                await _barberRepository.DeleteAll();
+            }
         }
-
         public async Task<Barber> Get(Expression<Func<Barber, bool>> filter)
         {
-            return await _barberRepository.Get(filter);
+            if (BarberValidator.BarberExists(_barberRepository.Get(filter).Result.Id))
+            {
+                return await _barberRepository.Get(filter);
+            }
+            else
+            {
+                throw new ArgumentException("Validation didn't pass.");
+            }
         }
-
         public async Task<IEnumerable<Barber>> GetAll()
         {
-            return await _barberRepository.GetAll();
+            if (await _barberRepository.GetCount() <= 0)
+            {
+                throw new ArgumentException("Nothing to get from here");
+            }
+            else
+            {
+                return await _barberRepository.GetAll();
+            }
         }
-
         public async Task<IEnumerable<Barber>> GetBarberBySpeciality(string speciality)
         {
             Expression<Func<Barber, bool>> filter = barber => barber.Speciality.Type == speciality;
-            return await _barberRepository.Find(filter);
+            if (BarberValidator.BarberExists(_barberRepository.Get(filter).Result.Id))
+            {
+                return await _barberRepository.Find(filter);
+            }
+            else
+            {
+                throw new ArgumentException("Validation didn't pass.");
+            }
         }
-
         public async Task RemoveRange(IEnumerable<Barber> entities)
         {
-            await _barberRepository.RemoveRange(entities);
+            if (entities.Count() < 0)
+            {
+                throw new ArgumentException("Validation didn't pass.");
+            }
+            else
+            {
+                await _barberRepository.RemoveRange(entities);
+            }
         }
-
-        public async Task UpdateByName(Guid id, Barber barber)
+        public async Task Update(Guid id, Barber barber)
         {
             Expression<Func<Barber, bool>> filter = barber => barber.Id == id;
-            Barber entity = _barberRepository.Get(filter).Result;
-            entity = barber;
-            await _barberRepository.Update(entity);
+            if (BarberValidator.BarberExists(_barberRepository.Get(filter).Result.Id))
+            {
+                Barber entity = _barberRepository.Get(filter).Result;
+                entity = barber;
+                await _barberRepository.Update(entity);
+            }
         }
     }
 }
